+++ packages/components/src/components/ScalarSearchResults/index.ts
@@ -1,0 +1,2 @@
+export { default as ScalarSearchResultList } from './ScalarSearchResultList.vue'
+export { default as ScalarSearchResultItem } from './ScalarSearchResultItem.vue'