+++ packages/scalar_fastapi/scalar_fastapi/__init__.py
@@ -1,0 +1,1 @@
+from .scalar_fastapi import get_scalar_api_reference