+++ packages/themes/vite.config.ts
@@ -14,6 +14,10 @@ export default defineConfig({
           dest: './',
         },
         {
+          src: 'src/fonts.css',
+          dest: './',
+        },
+        {
           src: 'src/scrollbar.css',
           dest: './',
         },
@@ -33,6 +37,8 @@ export default defineConfig({
     rollupOptions: {
       external: ['vue'],
     },
+    // Don't minify CSS so we can use it in stuff like the theme editor
+    cssMinify: false,
   },
   test: {
     coverage: {